# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: QR Restaurant API
  version: 1.0.0
  description: |
    REST API платформы цифрового обслуживания ресторана: QR-меню, заказы без регистрации,
    state machine заказа, кухонный терминал, панель официанта, вызов официанта через RabbitMQ.

    **Аутентификация**
    * Персонал (ADMIN, COOK, WAITER) — `Authorization: Bearer <JWT>` из `POST /auth/login`.
    * Гость — заголовок `X-Table-Token: <UUID из QR-кода стола>`. Регистрация не нужна.

    **Формат**: JSON (UTF-8), даты — ISO 8601 (UTC), деньги — число с 2 знаками.
    Ошибки: `{ "error": { "code": "...", "message": "...", "details": ... } }`.

    **Реальное время** (Socket.io, тот же порт):
    * namespace `/` — персонал, `auth: { token: <JWT> }`; события `ORDER_CREATED`, `ORDER_STATUS_CHANGED`,
      `NEW_WAITER_CALL`, `WAITER_CALL_RESOLVED`, `STOP_LIST_CHANGED`;
    * namespace `/guest` — гость, `auth: { tableToken: <UUID> }`; события своего стола
      `ORDER_STATUS_CHANGED`, `WAITER_CALL_RESOLVED`, а также `STOP_LIST_CHANGED`.
servers:
  - url: http://localhost:5000/api/v1
    description: Локальная разработка
...
paths:
...
  /menu:
    get:
      operationId: getMenu
      tags: [Menu]
      summary: Меню с категориями и блюдами
      description: 'Публичный. Блюда из стоп-листа возвращаются с `isAvailable: false`.'
      security: []
      parameters:
        - name: available
          in: query
          description: '`true` — только доступные блюда'
          schema: { type: string, enum: ['true', 'false'] }
        - name: restaurantId
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Меню
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MenuCategory' }
  /menu/categories:
    get:
      operationId: listCategories
      tags: [Menu]
      summary: Список категорий
      security: []
      responses:
        '200':
          description: Категории
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }
    post:
      operationId: createCategory
      tags: [Menu]
      summary: Создать категорию (ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryInput' }
            example: { name: Пицца, sortOrder: 1 }
      responses:
        '201':
          description: Создана
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /menu/categories/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch:
      operationId: updateCategory
      tags: [Menu]
      summary: Изменить категорию (ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryUpdate' }
      responses:
        '200':
          description: Изменена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteCategory
      tags: [Menu]
      summary: Удалить пустую категорию (ADMIN)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Удалена }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: В категории есть блюда (CATEGORY_NOT_EMPTY)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /menu/stop-list:
    get:
      operationId: getStopList
      tags: [Menu]
      summary: Блюда в стоп-листе (персонал)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Блюда с isAvailable = false
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Dish' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /menu/dishes:
    post:
      operationId: createDish
      tags: [Menu]
      summary: Создать блюдо (ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DishInput' }
            example:
              categoryId: 1
              name: Маргарита
              description: Томатный соус, моцарелла, базилик
              price: 590
              calories: 820
              ingredients: [Тесто, Томатный соус, Моцарелла, Базилик]
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dish' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /menu/dishes/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getDish
      tags: [Menu]
      summary: Карточка блюда
      security: []
      responses:
        '200':
          description: Блюдо
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dish' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateDish
      tags: [Menu]
      summary: Изменить блюдо (ADMIN)
      description: Новая цена действует только для новых заказов — созданные хранят снимок цены (BE-06).
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DishUpdate' }
            example: { price: 650 }
      responses:
        '200':
          description: Изменено
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dish' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteDish
      tags: [Menu]
      summary: Удалить блюдо (ADMIN)
      description: История заказов сохраняется благодаря снимкам названия и цены в позициях.
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Удалено }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /tables:
    get:
      operationId: listTables
      tags: [Tables]
      summary: Список столов с токенами (ADMIN)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Столы
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Table' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createTable
      tags: [Tables]
      summary: Создать стол (ADMIN)
      description: Генерируется tableToken — UUID v4 из криптостойкого генератора (BE-03).
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [number]
              properties:
                number: { type: string, maxLength: 20 }
                restaurantId: { type: integer }
            example: { number: '6' }
      responses:
        '201':
          description: Стол создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Table' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tables/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch:
      operationId: updateTable
      tags: [Tables]
      summary: Переименовать стол (ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [number]
              properties:
                number: { type: string }
      responses:
        '200':
          description: Изменён
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Table' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteTable
      tags: [Tables]
      summary: Удалить стол без заказов (ADMIN)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Удалён }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /tables/check/{token}:
    get:
      operationId: checkTableToken
      tags: [Tables]
      summary: Проверить QR-токен (гость)
      description: Публичный. Возвращает контекст стола для гостевой сессии (BE-04).
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Токен действителен
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TableSession' }
        '404':
          description: Токен не найден (INVALID_TABLE_TOKEN)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /orders/quote:
    post:
      operationId: quoteOrder
      tags: [Orders]
      summary: Расчёт корзины сервером (гость)
      description: Проверяет стоп-лист и считает итог по ценам из БД, заказ не создаёт (FE-02).
      security: [{ tableToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items: { $ref: '#/components/schemas/OrderItemInput' }
      responses:
        '200':
          description: Подтверждённый итог
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderQuote' }
        '400': { $ref: '#/components/responses/OrderRejected' }
        '401': { $ref: '#/components/responses/InvalidTableToken' }
  /orders:
    post:
      operationId: createOrder
      tags: [Orders]
      summary: Создать заказ (гость)
      description: |
        Сервер заново проверяет стоп-лист и цены (BE-05), фиксирует снимок цены каждой позиции
        и создаёт заказ в статусе ACCEPTED (BE-06). Цены из запроса клиента игнорируются.
        Если передан `expectedTotal` и он не совпал с серверной суммой — 409 PRICE_CHANGED.
      security: [{ tableToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrderRequest' }
            example:
              items: [{ dishId: 1, quantity: 2 }, { dishId: 10, quantity: 1 }]
              comment: Без лука
              expectedTotal: 1400
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400': { $ref: '#/components/responses/OrderRejected' }
        '401': { $ref: '#/components/responses/InvalidTableToken' }
        '409':
          description: Цены изменились (PRICE_CHANGED)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: PRICE_CHANGED
                  message: Цены в меню изменились — проверьте корзину и подтвердите заказ ещё раз
                  details: { expectedTotal: 1400, actualTotal: 1480 }
...
  /orders/active:
    get:
      operationId: getActiveOrders
      tags: [Orders]
      summary: Активные заказы — кухня и официант (COOK, WAITER, ADMIN)
      description: 'Очередь: старые заказы первыми. `availableTransitions` — кнопки, доступные роли.'
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          description: Фильтр через запятую, например `READY` или `ACCEPTED,COOKING`
          schema: { type: string }
      responses:
        '200':
          description: Заказы
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /orders/{id}/status:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch:
      operationId: changeOrderStatus
      tags: [Orders]
      summary: Сменить статус заказа (state machine)
      description: |
        | Переход | Роль |
        |---|---|
        | ACCEPTED → COOKING | COOK |
        | COOKING → READY | COOK |
        | READY → SERVED | WAITER |
        | SERVED → PAID | ADMIN |
        | ACCEPTED → CANCELLED | COOK, ADMIN |
        | COOKING / READY / SERVED → CANCELLED | ADMIN |

        Несуществующий переход — 400 INVALID_STATUS_TRANSITION; переход не для этой роли — 403
        TRANSITION_FORBIDDEN_FOR_ROLE; статус уже изменил другой сотрудник — 409 ORDER_STATUS_CONFLICT.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/OrderStatus' }
                reason: { type: string, maxLength: 500, description: 'Причина (например, для CANCELLED)' }
            example: { status: COOKING }
      responses:
        '200':
          description: Статус изменён
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400':
          description: Недопустимый переход
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: INVALID_STATUS_TRANSITION
                  message: 'Недопустимый переход статуса заказа: ACCEPTED → PAID'
                  details: { from: ACCEPTED, to: PAID, allowedTargets: [COOKING, CANCELLED] }
        '403':
          description: Переход запрещён для роли
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: TRANSITION_FORBIDDEN_FOR_ROLE
                  message: 'Ваша роль не может выполнить этот переход статуса: ACCEPTED → COOKING'
                  details: { from: ACCEPTED, to: COOKING, role: WAITER, allowedRoles: [COOK] }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT сотрудника из POST /auth/login
    tableToken:
      type: apiKey
      in: header
      name: X-Table-Token
      description: UUID стола из QR-кода
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
    ValidationError:
      description: Ошибка валидации (VALIDATION_ERROR)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: VALIDATION_ERROR
              message: Ошибка валидации входных данных
              details: [{ field: price, message: Цена должна быть больше 0 }]
    Unauthorized:
      description: Нет или неверный JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Недостаточно прав (FORBIDDEN)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Конфликт данных
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    InvalidTableToken:
      description: Нет или неверный X-Table-Token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: INVALID_TABLE_TOKEN, message: QR-код стола недействителен } }
    OrderRejected:
      description: Заказ отклонён — блюдо в стоп-листе или не найдено, некорректная корзина
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: DISH_IN_STOP_LIST
              message: 'Блюдо временно недоступно (стоп-лист): Тирамису'
              details: { dishes: [{ id: 9, name: Тирамису }] }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details: {}
...
    OrderStatus:
      type: string
      enum: [ACCEPTED, COOKING, READY, SERVED, PAID, CANCELLED]
...
    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        sortOrder: { type: integer }
        restaurantId: { type: integer }
    CategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 100 }
        sortOrder: { type: integer, minimum: 0 }
        restaurantId: { type: integer }
    CategoryUpdate:
      type: object
      properties:
        name: { type: string }
        sortOrder: { type: integer }
    Dish:
      type: object
      properties:
        id: { type: integer }
        categoryId: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        price: { type: number, example: 590 }
        calories: { type: integer, nullable: true, description: ккал }
        imageUrl: { type: string, nullable: true }
        isAvailable: { type: boolean, description: false — блюдо в стоп-листе }
        ingredients:
          type: array
          items: { type: string }
        updatedAt: { type: string, format: date-time }
    DishInput:
      type: object
      required: [categoryId, name, price]
      properties:
        categoryId: { type: integer }
        name: { type: string, maxLength: 150 }
        description: { type: string, nullable: true }
        price: { type: number, minimum: 0.01, multipleOf: 0.01 }
        calories: { type: integer, minimum: 0, nullable: true }
        imageUrl: { type: string, format: uri, nullable: true }
        ingredients:
          type: array
          maxItems: 50
          items: { type: string }
        isAvailable: { type: boolean }
    DishUpdate:
      allOf:
        - $ref: '#/components/schemas/DishInput'
      description: Любое подмножество полей DishInput (хотя бы одно)
    MenuCategory:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        sortOrder: { type: integer }
        dishes:
          type: array
          items: { $ref: '#/components/schemas/Dish' }
    Table:
      type: object
      properties:
        id: { type: integer }
        number: { type: string }
        restaurantId: { type: integer }
        tableToken: { type: string, format: uuid }
        qrUrl: { type: string, example: 'http://localhost:5173/table/1b2a731a-e307-48b4-8379-34485e6be9db' }
        createdAt: { type: string, format: date-time }
    TableSession:
      type: object
      properties:
        valid: { type: boolean }
        table:
          type: object
          properties:
            id: { type: integer }
            number: { type: string }
        restaurant:
          type: object
          properties:
            id: { type: integer }
            name: { type: string }
    OrderItemInput:
      type: object
      required: [dishId, quantity]
      properties:
        dishId: { type: integer }
        quantity: { type: integer, minimum: 1, maximum: 50 }
    CreateOrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items: { $ref: '#/components/schemas/OrderItemInput' }
        comment: { type: string, maxLength: 500 }
        expectedTotal: { type: number, description: 'Сумма, которую видел гость' }
    OrderQuote:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              dishId: { type: integer }
              dishName: { type: string }
              quantity: { type: integer }
              price: { type: number }
              lineTotal: { type: number }
        totalAmount: { type: number }
    OrderItem:
      type: object
      properties:
        id: { type: integer }
        dishId: { type: integer, nullable: true, description: null — блюдо удалено из меню }
        dishName: { type: string, description: Снимок названия }
        quantity: { type: integer }
        orderedPrice: { type: number, description: Снимок цены на момент заказа }
        lineTotal: { type: number }
    Order:
      type: object
      properties:
        id: { type: integer }
        tableId: { type: integer }
        tableNumber: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        totalAmount: { type: number }
        comment: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        waitingSeconds: { type: integer, description: Секунд с момента заказа }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        availableTransitions:
          type: array
          description: Только для персонала — статусы, доступные роли
          items: { $ref: '#/components/schemas/OrderStatus' }
...
security:
  - bearerAuth: []